\name{summary.csboost}

\alias{summary.csboost}

\title{Summarizing csboost fits}

\description{The function is a method for class \code{csboost} objects that summarizes their fit.}

\usage{
\method{summary}{csboost}(object, \dots)
}

\arguments{
\item{object}{an object of class "\code{csboost}". A result of a call to the \code{csboost} function.}
\item{...}{further arguments passed to or from other methods.}
}

\details{\code{summary.csboost} prints a summary of the "\code{csboost}" object.}

\references{...}

\author{Sebastiaan Höppner}


\seealso{\code{\link{csboost}}, \code{\link{summary}}}

\examples{
# Provide an example (TODO).
}
