\name{plot.csboost}

\alias{plot.csboost}

\title{Plot method for csboost fits}

\description{Plot the expected savings against the iteration index.}

\usage{
\method{plot}{csboost}(x, \dots)
}

\arguments{
\item{x}{a fitted object of class "\code{csboost}".}
\item{...}{further arguments passed to or from other methods.}
}

\details{\code{plot.csboost} plots the evolution of the expected savings.}

\references{...}

\author{Sebastiaan Höppner}


\seealso{\code{\link{csboost}}, \code{\link{predict}}}

\examples{
# Provide an example (TODO).
}
